refactor: move settings screen to features
